[REFACTOR/#21] Dto 패키지 변경
