+++ app/src/main/java/org/sopt/dosopttemplate/data/model/remote/model/response/auth/ResponseSignupDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.model.remote.model.dto.response.auth
+package org.sopt.dosopttemplate.data.model.remote.model.response.auth
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
