+++ app/src/main/java/org/sopt/dosopttemplate/data/repository/HomeRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package org.sopt.dosopttemplate.data.repository
 
 import org.sopt.dosopttemplate.data.model.remote.api.ServicePool
-import org.sopt.dosopttemplate.data.model.remote.model.dto.response.follower.ResponseFollowerDto
+import org.sopt.dosopttemplate.data.model.remote.model.response.follower.ResponseFollowerDto
 
 class HomeRepositoryImpl {
     suspend fun getFollowerList(pageNumber: Int): Result<List<ResponseFollowerDto.FollowerData>> =
