+++ app/src/main/java/org/sopt/dosopttemplate/data/service/FollowerService.kt
@@ -1,6 +1,6 @@
 package org.sopt.dosopttemplate.data.service
 
-import org.sopt.dosopttemplate.data.model.remote.model.dto.response.follower.ResponseFollowerDto
+import org.sopt.dosopttemplate.data.model.remote.model.response.follower.ResponseFollowerDto
 import retrofit2.http.GET
 import retrofit2.http.Query
